# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Enrollment System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Course {
  id: number;
  name: string;
  description: string;
  duration: string;
  fees: number;
  category: string;
  seats: number;
  createdAt: string;
}

export interface CourseInput {
  name: string;
  description: string;
  duration: string;
  fees: number;
  category: string;
  seats: number;
}

export interface CourseUpdate {
  name?: string;
  description?: string;
  duration?: string;
  fees?: number;
  category?: string;
  seats?: number;
}

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface Application {
  id: number;
  /** @nullable */
  userId?: string | null;
  studentName: string;
  email: string;
  phone: string;
  dob: string;
  address: string;
  courseId: number;
  /** @nullable */
  courseName?: string | null;
  status: ApplicationStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ApplicationInput {
  studentName: string;
  email: string;
  phone: string;
  dob: string;
  address: string;
  courseId: number;
  /** @nullable */
  documentName?: string | null;
}

export type ApplicationStatusUpdateStatus = typeof ApplicationStatusUpdateStatus[keyof typeof ApplicationStatusUpdateStatus];


export const ApplicationStatusUpdateStatus = {
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface ApplicationStatusUpdate {
  status: ApplicationStatusUpdateStatus;
  /** @nullable */
  notes?: string | null;
}

export interface AdminStats {
  totalApplications: number;
  pendingApplications: number;
  acceptedApplications: number;
  rejectedApplications: number;
  totalCourses: number;
  totalStudents: number;
}

export interface CourseReport {
  courseId: number;
  courseName: string;
  totalApplications: number;
  accepted: number;
  rejected: number;
  pending: number;
}

export interface ContactInput {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  message: string;
  createdAt: string;
}

export interface StudentProfile {
  id: number;
  userId: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  dob?: string | null;
  /** @nullable */
  address?: string | null;
  createdAt?: string;
}

export interface ProfileUpdate {
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  dob?: string | null;
  /** @nullable */
  address?: string | null;
}

export type ListCoursesParams = {
category?: string;
search?: string;
};

export type ListApplicationsParams = {
status?: ListApplicationsStatus;
courseId?: number;
};

export type ListApplicationsStatus = typeof ListApplicationsStatus[keyof typeof ListApplicationsStatus];


export const ListApplicationsStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;