# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: College Enrollment System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: courses
    description: Course management
  - name: applications
    description: Enrollment applications
  - name: admin
    description: Admin dashboard
  - name: contact
    description: Contact form
  - name: profile
    description: Student profile

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"

  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update course (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List all applications (admin)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, accepted, rejected]
        - in: query
          name: courseId
          schema:
            type: integer
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: submitApplication
      tags: [applications]
      summary: Submit an enrollment application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Submitted application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  /applications/my:
    get:
      operationId: getMyApplications
      tags: [applications]
      summary: Get current student's applications
      responses:
        "200":
          description: My applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"

  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  /applications/{id}/status:
    patch:
      operationId: updateApplicationStatus
      tags: [applications]
      summary: Accept or reject an application (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard statistics
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/reports:
    get:
      operationId: getAdminReports
      tags: [admin]
      summary: Get enrollment reports by course
      responses:
        "200":
          description: Report data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseReport"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Message submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current student profile
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Course:
      type: object
      required: [id, name, description, duration, fees, category, seats, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        duration:
          type: string
        fees:
          type: number
        category:
          type: string
        seats:
          type: integer
        createdAt:
          type: string

    CourseInput:
      type: object
      required: [name, description, duration, fees, category, seats]
      properties:
        name:
          type: string
        description:
          type: string
        duration:
          type: string
        fees:
          type: number
        category:
          type: string
        seats:
          type: integer

    CourseUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        duration:
          type: string
        fees:
          type: number
        category:
          type: string
        seats:
          type: integer

    Application:
      type: object
      required: [id, studentName, email, phone, dob, address, courseId, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["string", "null"]
        studentName:
          type: string
        email:
          type: string
        phone:
          type: string
        dob:
          type: string
        address:
          type: string
        courseId:
          type: integer
        courseName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    ApplicationInput:
      type: object
      required: [studentName, email, phone, dob, address, courseId]
      properties:
        studentName:
          type: string
        email:
          type: string
        phone:
          type: string
        dob:
          type: string
        address:
          type: string
        courseId:
          type: integer
        documentName:
          type: ["string", "null"]

    ApplicationStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, rejected]
        notes:
          type: ["string", "null"]

    AdminStats:
      type: object
      required: [totalApplications, pendingApplications, acceptedApplications, rejectedApplications, totalCourses, totalStudents]
      properties:
        totalApplications:
          type: integer
        pendingApplications:
          type: integer
        acceptedApplications:
          type: integer
        rejectedApplications:
          type: integer
        totalCourses:
          type: integer
        totalStudents:
          type: integer

    CourseReport:
      type: object
      required: [courseId, courseName, totalApplications, accepted, rejected, pending]
      properties:
        courseId:
          type: integer
        courseName:
          type: string
        totalApplications:
          type: integer
        accepted:
          type: integer
        rejected:
          type: integer
        pending:
          type: integer

    ContactInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        message:
          type: string

    ContactMessage:
      type: object
      required: [id, name, email, message, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        message:
          type: string
        createdAt:
          type: string

    StudentProfile:
      type: object
      required: [id, userId, name, email]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        dob:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string

    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        dob:
          type: ["string", "null"]
        address:
          type: ["string", "null"]