# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Enrollment System API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalApplications: number;
  pendingApplications: number;
  acceptedApplications: number;
  rejectedApplications: number;
  totalCourses: number;
  totalStudents: number;
}