# lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Enrollment System API
 * OpenAPI spec version: 0.1.0
 */

export interface ApplicationInput {
  studentName: string;
  email: string;
  phone: string;
  dob: string;
  address: string;
  courseId: number;
  /** @nullable */
  documentName?: string | null;
}