# lib/api-zod/src/generated/types/applicationStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Enrollment System API
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationStatusUpdateStatus } from './applicationStatusUpdateStatus';

export interface ApplicationStatusUpdate {
  status: ApplicationStatusUpdateStatus;
  /** @nullable */
  notes?: string | null;
}