# lib/api-zod/src/generated/types/courseUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Enrollment System API
 * OpenAPI spec version: 0.1.0
 */

export interface CourseUpdate {
  name?: string;
  description?: string;
  duration?: string;
  fees?: number;
  category?: string;
  seats?: number;
}