# lib/api-zod/src/generated/types/studentProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Enrollment System API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentProfile {
  id: number;
  userId: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  dob?: string | null;
  /** @nullable */
  address?: string | null;
  createdAt?: string;
}